extract Participant, use FlatList
